Fix in EDNS - type for max size.

+++ src/dnslib/edns.h
@@ -203,7 +203,7 @@ int dnslib_edns_has_option(const dnslib_opt_rr_t *opt_rr, uint16_t code);
  * \retval DNSLIB_ESPACE
  */
 short dnslib_edns_to_wire(const dnslib_opt_rr_t *opt_rr, uint8_t *wire,
-                          short max_size);
+                          size_t max_size);
 
 /*!
  * \brief Returns size of the OPT RR in wire format.
